Added expiration of posts, edit page, and completed new posting
