+++ backend/functions/make-posting.js
@@ -1,8 +1,8 @@
 const { createPost } = require('../utils')
 
-module.exports = async (user, data) => {
+module.exports = async (user, data, postId) => {
   if (!user) throw Error('not authorized')
   if (!data.title || !data.description || !data.contactInfo || !data.skills) throw Error('missing post parameters')
 
-  return createPost(user._id, data.title, data.description, data.contactInfo, data.skills)
+  return createPost(user._id, data.title, data.description, data.contactInfo, data.skills, postId)
 }
