+++ client/routes.jsx
@@ -8,6 +8,7 @@ import Dashboard from './dashboard/dashboard'
 import PageNotFound from './page-not-found/page-not-found'
 import Postings from './postings/postings'
 import MakePosting from './make-posting/make-posting'
+import EditPosting from './edit-posting/edit-posting'
 // import theme from './theme'
 
 //
@@ -20,6 +21,7 @@ export default (props) =>
       <Route component={SplashPage} path='/' exact />
       <Route component={Postings} path='/postings' exact />
       <Route component={MakePosting} path='/makeposting' exact />
+      <Route component={EditPosting} path='/editposting/:id' exact />
       <Route component={Dashboard} path='/dashboard' exact />
       <Route component={PageNotFound} />
     </Switch>
